# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v1
components:
  schemas:
    OrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        total:
          type: number
          format: double
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductDTO"
        userName:
          type: string
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
        name:
          type: string
        company:
          type: string
        price:
          type: number
          format: double
        description:
          type: string
        stock:
          type: integer
          format: int32
        isAccepted:
          type: boolean
        date:
          type: string
          format: date-time
        views_count:
          type: integer
          format: int32
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        imageBase64:
          type: string
    Review:
      type: object
      properties:
        reviewId:
          type: integer
          format: int64
        username:
          type: string
        comment:
          type: string
        calification:
          type: integer
          format: int32
        creationDate:
          type: string
          format: date-time
    ReviewDTO:
      type: object
      properties:
        reviewId:
          type: integer
          format: int64
        username:
          type: string
        comment:
          type: string
        calification:
          type: integer
          format: int32
        creationDate:
          type: string
          format: date
    UserCreationDTO:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: array
          items:
            type: string
        orders:
          type: array
          items:
            $ref: "#/components/schemas/OrderDTO"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ReviewDTO"
        cart:
          type: array
          items:
            $ref: "#/components/schemas/ProductDTO"
        cartPrice:
          type: number
          format: double
        historicalOrderPrices:
          type: array
          items:
            type: number
            format: double
        imageBase64:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: array
          items:
            type: string
        orders:
          type: array
          items:
            $ref: "#/components/schemas/OrderDTO"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ReviewDTO"
        cart:
          type: array
          items:
            $ref: "#/components/schemas/ProductDTO"
        cartPrice:
          type: number
          format: double
        historicalOrderPrices:
          type: array
          items:
            type: number
            format: double
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
servers:
  - url: https://localhost:8443
    description: Generated server url
paths:
  /api/v1/users/{id}/image:
    get:
      tags:
        - api-user-controller
      summary: Get user profile image
      description: Retrieves the profile image for the specified user
      operationId: getUserImage
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Numeric ID of the user
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: User image retrieved successfully
          content:
            image/*:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized - Missing or invalid authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found or has no profile image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - api-user-controller
      summary: Replace user profile image
      description: Uploads a new profile image replacing any existing one
      operationId: replaceUserImage
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Numeric ID of the user
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
                  description: The image file to upload (JPG/PNG)
              required:
                - imageFile
      responses:
        "200":
          description: Image replaced successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  imageUrl:
                    type: string
        "400":
          description: Invalid file format or missing file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - api-user-controller
      summary: Upload user profile image
      description: Uploads a profile image (requires no existing image)
      operationId: uploadUserImage
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Numeric ID of the user
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
                  description: The image file to upload (JPG/PNG)
              required:
                - imageFile
      responses:
        "201":
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  imageUrl:
                    type: string
        "400":
          description: User already has a profile image (use PUT instead)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
      - api-user-controller
      summary: Delete user image
      description: Delete a user's image.
      operationId: deleteProductImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User image deleted successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Image doesn't exist
          content:
            '*/*':
              schema:
                type: object
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
  /api/v1/users:
    get:
      tags:
      - api-user-controller
      summary: Get all users
      description: Retrieve a list of all users in the system.
      operationId: getAllUsers
      responses:
        "200":
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDTO"
    put:
      tags:
      - api-user-controller
      summary: Update user profile
      description: Update the details of an existing user.
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreationDTO"
        required: true
      responses:
        "200":
          description: Profile updated successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
      - api-user-controller
      summary: Create a new user
      description: Register a new user in the system.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreationDTO"
        required: true
      responses:
        "200":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "400":
          description: Username already exists
          content:
            '*/*':
              schema:
                type: object
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/products/{id}/image:
    get:
      tags:
      - api-product-controller
      summary: Get product image
      description: Retrieve the image of a specific product by its ID.
      operationId: getProductImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Product image retrieved successfully
          content:
            image/jpeg: {}
        "404":
          description: Product image not found
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
      - api-product-controller
      summary: Replace product image
      description: Replace the existing image of a product with a new one.
      operationId: replaceProductImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "200":
          description: Product image replaced successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    post:
      tags:
      - api-product-controller
      summary: Upload product image
      description: Upload a new image for a specific product.
      operationId: createProductImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "200":
          description: Product image uploaded successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    delete:
      tags:
      - api-product-controller
      summary: Delete product image
      description: Delete the image of a specific product.
      operationId: deleteProductImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Product image deleted successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
...
  /api/v1/users/{id}:
    get:
      tags:
      - api-user-controller
      summary: Get user by ID
      description: Retrieve a user by their unique ID.
      operationId: getUserById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - api-user-controller
      summary: Delete user
      description: Delete a user by their unique ID.
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User deleted successfully
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/users/profile:
    get:
      tags:
      - api-user-controller
      summary: Get user profile
      description: Retrieve the profile details of the authenticated user.
      operationId: getProfile
      responses:
        "200":
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
  /api/v1/users/by-username/{username}:
    get:
      tags:
      - api-user-controller
      summary: Get user ID by username
      description: Retrieve a user's ID by their username.
      operationId: getUserIdByUsername
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: User ID retrieved successfully
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: integer
                format: int64
...